+++ resources/lang/en/messages.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'email' => 'The :attribute must be a valid email address.',
+];