+++ src/lang/en/validate.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-  'email-check' => 'Invalid Email Address'
-];