+++ tests/EmailCheckerTest.php
@@ -1,40 +1,24 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: tintnaingwin
- * Date: 4/17/18
- * Time: 2:47 PM
- */
-use Tintnaingwin\EmailChecker\EmailChecker;
-use PHPUnit\Framework\TestCase;
+
+namespace Tintnaingwin\EmailChecker\Tests;
+
+use Tintnaingwin\EmailChecker\Facades\EmailChecker;
 
 class EmailCheckerTest extends TestCase
 {
     public $email_checker;
 
-    /**
-     * EmailCheckerTest constructor.
-     */
-
-    public function __construct($name = null, array $data = [], $dataName = '')
+    /** @test */
+    public function it_can_validate_email_address()
     {
-        parent::__construct($name, $data, $dataName);
-        $this->email_checker = new EmailChecker();
+        $this->assertTrue(EmailChecker::check('amigo.k8@gmail.com'));
+        $this->assertFalse(EmailChecker::check('example@example.com'));
     }
 
-    public function testEmailCheckerIsTrue()
+    /** @test */
+    public function it_will_fail_if_email_address_is_disposable_mail()
     {
-        $this->assertTrue($this->email_checker->check('amigo.k8@gmail.com'));
-    }
-
-    public function testEmailCheckerIsFalse()
-    {
-        $this->assertFalse($this->email_checker->check('example@example.com'));
-    }
-
-    public function testDisposableMailIsFalse()
-    {
-        $this->assertFalse($this->email_checker->check('amigo.k8@0-mail.com'));
+        $this->assertFalse(EmailChecker::check('amigo.k8@0-mail.com'));
     }
 
 }