+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## v2.0.0 (2019-11-12)
+### Added
+- Added Validation Rule Object
+
+### Changes
+- The EmailChecker.php have been moved to the new tintnaingwin/email-checker-php package.
+
+### Refactoring
+- Code refactoring
+
 ## v1.0.0 (2018-04-18)
 
 ### Added
@@ -11,4 +21,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Laravel Package Discovery
 
 ### Fixed
-- Fix the each() function is deprecated in php 7.2 .
+- Fix the each() function is deprecated in php 7.2.